extract common code into BackfillBase
